Implementando autenticação e autorização.

+++ backend/Services/TeamService/TeamService.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using backend.Data;
 using backend.Dtos.Team;
 using backend.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace backend.Services.TeamService
@@ -14,8 +16,10 @@ namespace backend.Services.TeamService
     {
         private readonly IMapper _mapper;
         private readonly DataContext _context;
-        public TeamService(IMapper mapper, DataContext context)
+        private readonly IHttpContextAccessor _httpContextAccessor ;
+        public TeamService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
+            this._httpContextAccessor  = httpContextAccessor;
             this._context = context;
             this._mapper = mapper;
 
@@ -25,10 +29,12 @@ namespace backend.Services.TeamService
         {
             ServiceResponse<List<GetTeamDto>> serviceResponse = new ServiceResponse<List<GetTeamDto>>();
             Team team = _mapper.Map<Team>(newTeam);
+            team.Player = await _context.Players.FirstOrDefaultAsync(u => u.Id == GetUserId());
             
             await _context.Teams.AddAsync(team);
             await _context.SaveChangesAsync();
-            serviceResponse.Data = (_context.Teams.Select(p => _mapper.Map<GetTeamDto>(p))).ToList();
+            
+            serviceResponse.Data = (_context.Teams.Where(t => t.Player.Id == GetUserId()).Select(p => _mapper.Map<GetTeamDto>(p))).ToList();
             return serviceResponse;
         }
 
@@ -36,7 +42,7 @@ namespace backend.Services.TeamService
         {
             ServiceResponse<List<GetTeamDto>> serviceResponse = new ServiceResponse<List<GetTeamDto>>();
 
-            List<Team> dbTeams = await _context.Teams.ToListAsync();
+            List<Team> dbTeams = await _context.Teams.Where(t => t.Player.Id == GetUserId()).ToListAsync();
             serviceResponse.Data = dbTeams.Select(p => _mapper.Map<GetTeamDto>(p)).ToList();
             return serviceResponse;
         }
@@ -45,7 +51,7 @@ namespace backend.Services.TeamService
         {
             ServiceResponse<GetTeamDto> serviceResponse = new ServiceResponse<GetTeamDto>();
 
-            Team dbTeam = await _context.Teams.FirstOrDefaultAsync(c => c.Id == id);
+            Team dbTeam = await _context.Teams.FirstOrDefaultAsync(c => c.Id == id && c.Player.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetTeamDto>(dbTeam);
             return serviceResponse;
         }
@@ -55,11 +61,18 @@ namespace backend.Services.TeamService
             ServiceResponse<List<GetTeamDto>> serviceResponse = new ServiceResponse<List<GetTeamDto>>();
             try
             {
-                Team team = await _context.Teams.FirstAsync(c => c.Id == id);
-                _context.Teams.Remove(team);
-                await _context.SaveChangesAsync();
-                
-                serviceResponse.Data = (_context.Teams.Select(c => _mapper.Map<GetTeamDto>(c))).ToList();
+                Team team = await _context.Teams.FirstAsync(c => c.Id == id && c.Player.Id == GetUserId());
+                if(team != null){
+                    _context.Teams.Remove(team);
+                    await _context.SaveChangesAsync();
+
+                    serviceResponse.Data = (_context.Teams.Where(c => c.Player.Id == GetUserId()).Select(c => _mapper.Map<GetTeamDto>(c))).ToList();
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found.";
+                }
             }
             catch (Exception ex)
             {
@@ -69,11 +82,25 @@ namespace backend.Services.TeamService
             return serviceResponse;
         }
 
-        /*
-            --update
-            Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
-            _context.Characters.Update(character);
-            await _context.SaveChangesAsync();
-        */
+        public async Task<ServiceResponse<GetTeamDto>> UpdateTeam(UpdateTeamDto updateTeam)
+        {
+            ServiceResponse<GetTeamDto> serviceResponse = new ServiceResponse<GetTeamDto>();
+
+            try
+            {
+                Team team = await _context.Teams.Include(c => c.Player).FirstOrDefaultAsync(c => c.Id == updateTeam.Id);
+                team.Name = updateTeam.Name;
+                _context.Teams.Update(team);
+                await _context.SaveChangesAsync();
+            }
+            catch (Exception ex)
+            {
+                serviceResponse.Success = false;
+                serviceResponse.Message = ex.Message;
+            }
+            return serviceResponse;
+        }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
     }
 }
